+++ apps/api/src/app/portfolio/portfolio.controller.ts
@@ -565,6 +565,7 @@ export class PortfolioController {
         [
           'currentNetWorth',
           'currentValueInBaseCurrency',
+          'dividendInBaseCurrency',
           'grossPerformance',
           'grossPerformanceWithCurrencyEffect',
           'netPerformance',