+++ apps/api/src/app/portfolio/types/performance-percentages.type.ts
@@ -3,6 +3,7 @@ import { HistoricalDataItem } from '@ghostfolio/common/interfaces';
 export type PerformancePercentages = Required<
   Pick<
     HistoricalDataItem,
+    | 'dividendInPercentageWithCurrencyEffect'
     | 'netPerformanceInPercentage'
     | 'netPerformanceInPercentageWithCurrencyEffect'
   >