+++ libs/common/src/lib/interfaces/portfolio-performance.interface.ts
@@ -2,6 +2,8 @@ export interface PortfolioPerformance {
   annualizedPerformancePercent?: number;
   currentNetWorth?: number;
   currentValueInBaseCurrency: number;
+  dividendInBaseCurrency: number;
+  dividendPercentageWithCurrencyEffect: number;
   netPerformance: number;
   netPerformancePercentage: number;
   netPerformancePercentageWithCurrencyEffect: number;